Moved tests for deployer and scheduler
