+++ tests/unit/managers/__init__.py
@@ -1,0 +1,1 @@
+"""kytos/maintenance.managers unit tests"""
